refactorin complete, waiting to test the functionality
